+++ src/main/java/de/caritas/cob/messageservice/api/model/rocket/chat/login/EmailsDTO.java
@@ -1,0 +1,17 @@
+package de.caritas.cob.messageservice.api.model.rocket.chat.login;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ * EmailsDTO for LoginResponseDTO
+ *
+ */
+@Getter
+@Setter
+@NoArgsConstructor
+public class EmailsDTO {
+  private String address;
+  private boolean verified;
+}