+++ src/test/java/de/caritas/cob/messageservice/testhelper/RocketChatFieldConstants.java
@@ -1,0 +1,11 @@
+package de.caritas.cob.messageservice.testhelper;
+
+public class RocketChatFieldConstants {
+
+  public static final String FIELD_NAME_RC_POST_GROUP_MESSAGES_READ = "rcPostGroupMessagesRead";
+  public static final String FIELD_VALUE_RC_POST_GROUP_MESSAGES_READ =
+      "/api/v1/subscriptions" + ".read";
+  public static final String FIELD_NAME_RC_GET_GROUP_INFO_URL = "rcGetGroupInfoUrl";
+  public static final String FIELD_VALUE_RC_GET_GROUP_INFO_URL =
+      "/api/v1/groups.info?roomId=qGTGZwAWjPhyfREya";
+}