+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.2.1](https://github.com/CaritasDeutschland/caritas-onlineBeratung-messageService/compare/v1.2.0...v1.2.1) (2020-08-26)
+
+## [1.1.0](https://github.com/CaritasDeutschland/caritas-onlineBeratung-messageService/compare/e094b600cc4512e2f107094be174f5fa84cbf581...v1.1.0) (2020-07-29)
+
+
+### Bug Fixes
+
+* added npm install for github release action ([e094b60](https://github.com/CaritasDeutschland/caritas-onlineBeratung-messageService/commit/e094b600cc4512e2f107094be174f5fa84cbf581))
+
+## 1.2.0 (2020-08-26)
+
+
+### Features
+
+* check on writing to a feedback chat if provided group id is valid ([8291bb2](https://github.com/CaritasDeutschland/caritas-onlineBeratung-messageService/commit/8291bb2e8596e4490f674ed7b4a6aded52f0141d))
+* Initial Commit ([9083810](https://github.com/CaritasDeutschland/caritas-onlineBeratung-messageService/commit/90838106445c029f43afc05307b8816888b0be7f))
+
+
+### Bug Fixes
+
+* logservice tests to be independent of logback rule ([87905bc](https://github.com/CaritasDeutschland/caritas-onlineBeratung-messageService/commit/87905bc837c35828b69c8b4bc93a75e35aa230d9))
+* removed unused field setter ([f98594b](https://github.com/CaritasDeutschland/caritas-onlineBeratung-messageService/commit/f98594bfede8b50e0404c44edc6a8a015caa4ea1))
+
+
 ## 1.1.0 (2020-07-29)